Feat: Add trigger logs
